# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/products":{"get":{"operationId":"ProductsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["products"]}},"/":{"get":{"operationId":"MainController_getMainPage","parameters":[],"responses":{"200":{"description":""}},"tags":["main (pages)"]}},"/signin":{"get":{"operationId":"MainController_getSignInPage","parameters":[],"responses":{"200":{"description":""}},"tags":["main (pages)"]}},"/signup":{"get":{"operationId":"MainController_getSignUpPage","parameters":[],"responses":{"200":{"description":""}},"tags":["main (pages)"]}},"/products/{name}":{"get":{"operationId":"MainController_getProductsPage","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["main (pages)"]}},"/product-comments/{productName}":{"get":{"operationId":"MainController_findAllByProduct","parameters":[{"name":"productName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["main (pages)"]}},"/cart":{"get":{"operationId":"MainController_getCartPage","parameters":[],"responses":{"200":{"description":""}},"tags":["main (pages)"]}},"/admin":{"get":{"operationId":"MainController_getAdminPage","parameters":[],"responses":{"200":{"description":""}},"tags":["main (pages)"]}},"/admin/users-create":{"get":{"operationId":"MainController_getAdminUsersCreatePage","parameters":[],"responses":{"200":{"description":""}},"tags":["main (pages)"]}},"/admin/products-create":{"get":{"operationId":"MainController_getAdminProductsPage","parameters":[],"responses":{"200":{"description":""}},"tags":["main (pages)"]}},"/api/v1/cart":{"post":{"operationId":"CartController_addToCart","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddToCartDto"}}}},"responses":{"201":{"description":""}},"tags":["cart"]}},"/api/v1/cart/{id}":{"delete":{"operationId":"CartController_removeFromCart","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["cart"]}},"/api/v1/cart/all":{"delete":{"operationId":"CartController_clearCart","parameters":[],"responses":{"200":{"description":""}},"tags":["cart"]}},"/api/v1/product-comments":{"post":{"operationId":"ProductCommentController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductCommentDto"}}}},"responses":{"201":{"description":""}},"tags":["product-comments"]}},"/api/v1/product-comments/{id}":{"patch":{"operationId":"ProductCommentController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductCommentDto"}}}},"responses":{"200":{"description":""}},"tags":["product-comments"]},"delete":{"operationId":"ProductCommentController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["product-comments"]}},"/api/v1/auth/signIn":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/v1/auth/signUp":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/v1/auth/profile":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/v1/admin/users":{"get":{"operationId":"AdminController_findAllUsers","parameters":[],"responses":{"200":{"description":""}},"tags":["admin"]},"post":{"operationId":"AdminController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["admin"]}},"/api/v1/admin/users/{id}":{"patch":{"operationId":"AdminController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["admin"]},"delete":{"operationId":"AdminController_removeUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["admin"]}},"/api/v1/admin/products":{"post":{"operationId":"AdminController_createProduct","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":""}},"tags":["admin"]}},"/api/v1/admin/products/{id}":{"patch":{"operationId":"AdminController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":""}},"tags":["admin"]},"delete":{"operationId":"AdminController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["admin"]}},"/api/v1/orders":{"post":{"operationId":"OrdersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["orders"]},"get":{"operationId":"OrdersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["orders"]}},"/api/v1/orders/{id}":{"get":{"operationId":"OrdersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["orders"]},"patch":{"operationId":"OrdersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrderDto"}}}},"responses":{"200":{"description":""}},"tags":["orders"]},"delete":{"operationId":"OrdersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["orders"]}}},"info":{"title":"BSE API","description":"BSE API description","version":"1.0","contact":{}},"tags":[{"name":"bse","description":""}],"servers":[],"components":{"schemas":{"AddToCartDto":{"type":"object","properties":{}},"CreateProductCommentDto":{"type":"object","properties":{}},"UpdateProductCommentDto":{"type":"object","properties":{}},"SignInDto":{"type":"object","properties":{}},"SignupDto":{"type":"object","properties":{}},"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"CreateProductDto":{"type":"object","properties":{}},"UpdateProductDto":{"type":"object","properties":{}},"CreateOrderDto":{"type":"object","properties":{}},"UpdateOrderDto":{"type":"object","properties":{}}}}}